requirements.txt
altair==5.4.1
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
extra-streamlit-components==0.1.71
fonttools==4.54.1
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.12.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
peewee==3.17.7
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
protobuf==5.28.3
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.39.0
streamlit-authenticator==0.4.1
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
yfinance==0.2.48
